requirements.txt
PyPDF2
langchain
openai
streamlit
python-dotenv
langchain-openai
python-docx
fpdf